--- src/components/Cart/CartBillingForm/index.ts
@@ -1,1 +1,0 @@
-export * from './CartBillingForm';
