+++ src/components/Cart/CartBillingInfo/index.ts
@@ -1,0 +1,1 @@
+export * from './CartBillingInfo';
