+++ src/components/Cart/CartConfirmations/CartConfirmations.module.scss
@@ -1,19 +1,6 @@
 @import '@style/_mixins';
 @import '@style/_media';
 
-.form {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 64px;
-}
-
-.form_grid {
-  display: grid;
-  grid-template: repeat(4, 1fr) / repeat(2, minmax(250px, auto));
-  gap: 32px;
-}
-
 .checkboxes {
   display: flex;
   flex-direction: column;
@@ -43,14 +30,4 @@
 
 .marketing {
   text-wrap: wrap;
-}
-
-@include mobile {
-  .submit {
-    display: flex;
-    flex-direction: column;
-  }
-  .form_grid {
-    grid-template: repeat(8, 1fr) / repeat(1, minmax(280px, auto));
-  }
 }
