+++ src/components/Cart/CartConfirmations/index.ts
@@ -1,0 +1,1 @@
+export * from './CartConfirmations';
