+++ src/types/cart/cartTypes.ts
@@ -1,0 +1,6 @@
+export interface BillingInputProps {
+  value: string;
+  error: string;
+  autocomplete: string[];
+  readonly: boolean;
+}
